examples/python/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
pillow==10.4.0
urllib3==2.2.2
websockets==13.0.1
